IActionResult, Action Overloading, Query Parameter, URL Parameter

+++ HotelFinder.API/Controllers/HotelsController.cs
@@ -15,7 +15,7 @@ namespace HotelFinder.API.Controllers
     public class HotelsController : Controller
     {
 
-        private IHotelService _hotelService;
+        private readonly IHotelService _hotelService;
 
         public HotelsController(IHotelService hotelService)
         {
@@ -23,33 +23,68 @@ namespace HotelFinder.API.Controllers
         }
 
 
-        public List<Hotel> Hotels()
+        public IActionResult Hotels()
         {
-            return _hotelService.GetAllHotels();
+            var hotels =  _hotelService.GetAllHotels();
+            return Ok(hotels); // 200 + data
         }
 
-        [HttpGet("{id}")]
-        public Hotel Hotel(int id)
+        [HttpGet]
+        [Route("hotelById/{id}")]
+        public IActionResult Hotel(int id)
         {
-            return _hotelService.GetHotelById(id);
+            var hotel = _hotelService.GetHotelById(id);
+            if (hotel == null)
+                return NotFound(); // 404
+
+            return Ok(hotel); // 200 + data
         }
+
+        [HttpGet]
+        [Route("[action]/{name}")]
+        public IActionResult HotelByName(string name)
+        {
+            return Ok();
+        }
+
+        [HttpGet]
+        [Route("[action]")]
+        public IActionResult HotelByIdAndName(int id, string name)
+        {
+            return Ok();
+        }
+
+
 
         [HttpPost]
-        public Hotel AddHotel([FromBody] Hotel hotel)
+        public IActionResult AddHotel([FromBody] Hotel hotel)
         {
-            return _hotelService.AddHotel(hotel);
+            if (string.IsNullOrEmpty(hotel.Name)
+                || string.IsNullOrEmpty(hotel.City))
+                return BadRequest("Empty Name or City"); // 400
+
+            return CreatedAtAction("Post", hotel); // 201 + data
         }
 
         [HttpPut]
-        public Hotel UpdateHotel([FromBody] Hotel hotel)
+        public IActionResult UpdateHotel([FromBody] Hotel hotel)
         {
-            return _hotelService.UpdateHotel(hotel);
+            var hotelFound = _hotelService.GetHotelById(hotel.Id);
+            if (hotelFound == null)
+                return NotFound(); //404
+
+            return Ok(_hotelService.UpdateHotel(hotel)); // 200 + data
         }
 
         [HttpDelete("{id}")]
-        public void DeleteHotel(int id)
+        public IActionResult DeleteHotel(int id)
         {
+            var hotelFound = _hotelService.GetHotelById(id);
+            if (hotelFound == null)
+                return NotFound(); //404
+
             _hotelService.DeleteHotel(id);
+            return Ok(); // 200
         }
